refactor: fully complete state re-architecting
